% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_code.R
\name{create_preamble_tex}
\alias{create_preamble_tex}
\title{Create preamble .tex for code font}
\usage{
create_preamble_tex(font_size = 12, monofont_path = get_fira_path())
}
\arguments{
\item{font_size}{a numeric. The font size, only available in pdf format.}

\item{monofont_path}{a string. The path to the font used to render code
chunks. It should link to a \code{.ttf} file. Only available in pdf format.}
}
\value{
a vector of characters representing an Rmd file (each element being
a line)
}
\description{
Create preamble .tex for code font
}
\examples{

create_preamble_tex()
create_preamble_tex(font_size = 18.0)

}
