% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_code.R
\name{generate_chunk_header}
\alias{generate_chunk_header}
\title{Generate R chunk header}
\usage{
generate_chunk_header(...)
}
\arguments{
\item{...}{The different options for R code chunks.}
}
\value{
a string of length 1
}
\description{
Function to generate R chunk header for rmarkdown rendering in different
output
}
\details{
To get the list of all accepted options, you can call
\code{names(knitr::opts_chunk$get())} and to get the default values you can
call \code{knitr::opts_chunk$get()}.

More information in the function #' \code{\link[knitr]{opts_chunk}} or
directly \url{https://yihui.org/knitr/options/#chunk-options} to see all
available options and their descriptions.
}
\examples{

generate_chunk_header()
generate_chunk_header(eval = TRUE, echo = TRUE)
generate_chunk_header(results = "asis")
generate_chunk_header(fig.width = "4px", fig.height = "3px")

}
