% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_file.R
\name{get_data}
\alias{get_data}
\title{Lecture des fichiers de donnée}
\usage{
get_data(path, erase_zero = TRUE, sep = ";", dec = ".", ...)
}
\arguments{
\item{path}{Chaîne de caractère. Chemin du fichier csv.}

\item{erase_zero}{Booléen. Remplace les valeurs nulle par \code{NA_real_}
(par défaut : \code{TRUE}).}

\item{sep}{Chaine de caractère. Séparateur de colonne (par défaut :
\code{;}).}

\item{dec}{Chaine de caractère. Séparateur décimal (par défaut : \code{.}).}

\item{...}{Autres paramètres de la fonction \code{read.csv}.}
}
\value{
Un objet de type \code{data.frame}.
}
\description{
Fonction de lecture de fichiers csv
}
\details{
par defaul
}
\examples{
file_path <- system.file("extdata", "file.csv", package = "TBox")
file_path <- normalizePath(file_path, winslash = "/", mustWork = FALSE)

df <- get_data(file_path, erase_zero = FALSE)

}
