% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_code.R
\name{get_fira_path}
\alias{get_fira_path}
\title{The path to the font Fira Code}
\usage{
get_fira_path()
}
\value{
a character vector of length 1 representing the path
}
\description{
This function returns the path to the font Fira Code installed with
the package TBox.
}
\details{
This function helps the other functions to find the path to the font Fira
Code to render documents and use this font by default for code chunks.
}
\examples{
get_fira_path()

}
