% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_code.R
\name{get_word_template_path}
\alias{get_word_template_path}
\title{The path to the word template}
\usage{
get_word_template_path()
}
\value{
a character vector of length 1 representing the path
}
\description{
This function returns the path to the word template installed with
the package TBox.
}
\details{
This function helps the other functions to find the template of the word
document used to render in \code{.docx} output.
}
\examples{
get_word_template_path()

}
