% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_file.R
\name{write_data}
\alias{write_data}
\title{Ecriture des fichiers de donnée}
\usage{
write_data(data, path)
}
\arguments{
\item{data}{\code{data.frame}. Données à écrire dans un fichier csv.}

\item{path}{Chaîne de caractère. Chemin du fichier csv.}
}
\value{
Chaîne de caractère. Chemin du fichier écrit.
}
\description{
Ecriture des fichiers de donnée
}
\details{
La fonction écrit les données dans un fichier csv avec les paramètres
suivants :
- séparateur de colonne : \code{;}
- séparateur décimal : \code{.}
- pas de guillemets
- pas de noms de lignes
- les \code{NA} sont remplacés par une chaîne vide.
}
\examples{
new_path <- tempfile("new_file", fileext = ".csv")
write_data(data = mtcars, path = new_path)

}
